Split ssolve into a library and application
Necessary for #5.
