+++ .clangd
@@ -1,3 +1,4 @@
 CompileFlags:
     Add:
         - -std=c11
+        - -Iinclude
