+++ include/ssolve/check.h
