+++ include/ssolve/puzzle.h
