+++ tests/stub.c
@@ -1,3 +1,5 @@
+#include "ssolve/puzzle.h"
+
 #include <stdio.h>
 
 int main(void)
